% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordinateSystem.R
\docType{class}
\name{CoordinateSystem}
\alias{CoordinateSystem}
\title{Coordinate system}
\description{
This class implements the coordinate system class.
}
\section{Super class}{
\code{\link[geoAPI:IdentifiedObject]{geoAPI::IdentifiedObject}} -> \code{CoordinateSystem}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoordinateSystem-new}{\code{CoordinateSystem$new()}}
\item \href{#method-CoordinateSystem-print}{\code{CoordinateSystem$print()}}
\item \href{#method-CoordinateSystem-getDimension}{\code{CoordinateSystem$getDimension()}}
\item \href{#method-CoordinateSystem-getAxis}{\code{CoordinateSystem$getAxis()}}
\item \href{#method-CoordinateSystem-clone}{\code{CoordinateSystem$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="geoAPI" data-topic="IdentifiedObject" data-id="getAlias"><a href='../../geoAPI/html/IdentifiedObject.html#method-IdentifiedObject-getAlias'><code>geoAPI::IdentifiedObject$getAlias()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geoAPI" data-topic="IdentifiedObject" data-id="getIdentifiers"><a href='../../geoAPI/html/IdentifiedObject.html#method-IdentifiedObject-getIdentifiers'><code>geoAPI::IdentifiedObject$getIdentifiers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geoAPI" data-topic="IdentifiedObject" data-id="getName"><a href='../../geoAPI/html/IdentifiedObject.html#method-IdentifiedObject-getName'><code>geoAPI::IdentifiedObject$getName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geoAPI" data-topic="IdentifiedObject" data-id="getRemarks"><a href='../../geoAPI/html/IdentifiedObject.html#method-IdentifiedObject-getRemarks'><code>geoAPI::IdentifiedObject$getRemarks()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geoAPI" data-topic="IdentifiedObject" data-id="setAlias"><a href='../../geoAPI/html/IdentifiedObject.html#method-IdentifiedObject-setAlias'><code>geoAPI::IdentifiedObject$setAlias()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geoAPI" data-topic="IdentifiedObject" data-id="setIdentifiers"><a href='../../geoAPI/html/IdentifiedObject.html#method-IdentifiedObject-setIdentifiers'><code>geoAPI::IdentifiedObject$setIdentifiers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geoAPI" data-topic="IdentifiedObject" data-id="setRemarks"><a href='../../geoAPI/html/IdentifiedObject.html#method-IdentifiedObject-setRemarks'><code>geoAPI::IdentifiedObject$setRemarks()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinateSystem-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinateSystem-new}{}}}
\subsection{Method \code{new()}}{
Create a new coordinate system. The CS must have at least
one axis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinateSystem$new(name, axes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character string. Name of the coordinate system.}

\item{\code{axes}}{A \code{list} of instances of \link{CoordinateSystemAxis}. There must
be at least one axis in the list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of \code{CoordinateSystem} or an error.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinateSystem-print"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinateSystem-print}{}}}
\subsection{Method \code{print()}}{
Print a summary of the coordinate system to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinateSystem$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinateSystem-getDimension"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinateSystem-getDimension}{}}}
\subsection{Method \code{getDimension()}}{
Retrieve the dimension of the coordinate system, the number
of axes that make up the coordinate system.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinateSystem$getDimension()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer value with the dimension of the coordinate system.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinateSystem-getAxis"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinateSystem-getAxis}{}}}
\subsection{Method \code{getAxis()}}{
Retrieve an axis of the coordinate system by its ordinal
number.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinateSystem$getAxis(dimension)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dimension}}{The dimension whose axis to retrieve. Note that this is
1-based.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \link{CoordinateSystemAxis} that forms the \code{dimension} of the
coordinate system, or an error if the \code{dimension} argument in deficient.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinateSystem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinateSystem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinateSystem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
