% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifiedObject.R
\docType{class}
\name{IdentifiedObject}
\alias{IdentifiedObject}
\title{Identified object}
\description{
This class is ancestor to many other, more useful classes. This
class just provides basic identification properties. It is not very useful to
instantiate this class directly; instead, use descendant classes.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IdentifiedObject-new}{\code{IdentifiedObject$new()}}
\item \href{#method-IdentifiedObject-getName}{\code{IdentifiedObject$getName()}}
\item \href{#method-IdentifiedObject-setAlias}{\code{IdentifiedObject$setAlias()}}
\item \href{#method-IdentifiedObject-getAlias}{\code{IdentifiedObject$getAlias()}}
\item \href{#method-IdentifiedObject-setIdentifiers}{\code{IdentifiedObject$setIdentifiers()}}
\item \href{#method-IdentifiedObject-getIdentifiers}{\code{IdentifiedObject$getIdentifiers()}}
\item \href{#method-IdentifiedObject-setRemarks}{\code{IdentifiedObject$setRemarks()}}
\item \href{#method-IdentifiedObject-getRemarks}{\code{IdentifiedObject$getRemarks()}}
\item \href{#method-IdentifiedObject-clone}{\code{IdentifiedObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiedObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiedObject-new}{}}}
\subsection{Method \code{new()}}{
Create a new object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiedObject$new(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character string. Name of the object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the object, or an error if the object could not
be created.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiedObject-getName"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiedObject-getName}{}}}
\subsection{Method \code{getName()}}{
Retrieve the name of the object. This method is mandatory in
the OGC standard.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiedObject$getName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string with the name of the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiedObject-setAlias"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiedObject-setAlias}{}}}
\subsection{Method \code{setAlias()}}{
Set aliases for the object. The aliases are added at the end
of the defined aliases.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiedObject$setAlias(alias)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alias}}{Character vector with aliases for the object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiedObject-getAlias"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiedObject-getAlias}{}}}
\subsection{Method \code{getAlias()}}{
Retrieve the list of aliases for the name of the object.
This method is optional in the OGC standard.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiedObject$getAlias()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of aliases for the name of the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiedObject-setIdentifiers"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiedObject-setIdentifiers}{}}}
\subsection{Method \code{setIdentifiers()}}{
Set identifiers for the object. The identifiers are added at the end
of the defined identifiers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiedObject$setIdentifiers(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Character vector with identifiers for the object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiedObject-getIdentifiers"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiedObject-getIdentifiers}{}}}
\subsection{Method \code{getIdentifiers()}}{
Retrieve the list of identifiers of the object. This method
is optional in the OGC standard.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiedObject$getIdentifiers()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of identifiers for the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiedObject-setRemarks"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiedObject-setRemarks}{}}}
\subsection{Method \code{setRemarks()}}{
Set remarks for the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiedObject$setRemarks(remarks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{remarks}}{Character string with remarks for the object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiedObject-getRemarks"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiedObject-getRemarks}{}}}
\subsection{Method \code{getRemarks()}}{
Retrieve the remarks of the object. This method is optional
in the OGC standard.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiedObject$getRemarks()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string with remarks or \code{NA} if no remarks have been
set.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IdentifiedObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IdentifiedObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IdentifiedObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
