% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordinateSystemAxis.R
\docType{class}
\name{CoordinateSystemAxis}
\alias{CoordinateSystemAxis}
\title{Coordinate system axis}
\description{
This class implements the coordinate system axis class. The
axis is always associated with a coordinate system.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoordinateSystemAxis-new}{\code{CoordinateSystemAxis$new()}}
\item \href{#method-CoordinateSystemAxis-print}{\code{CoordinateSystemAxis$print()}}
\item \href{#method-CoordinateSystemAxis-clone}{\code{CoordinateSystemAxis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinateSystemAxis-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinateSystemAxis-new}{}}}
\subsection{Method \code{new()}}{
Create a new axis instance for use in a coordinate system.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinateSystemAxis$new(name, abbreviation, direction, unit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name}

\item{\code{abbreviation}}{Abbreviation}

\item{\code{direction}}{Direction}

\item{\code{unit}}{Unit}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of this class or an error.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinateSystemAxis-print"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinateSystemAxis-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinateSystemAxis$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoordinateSystemAxis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoordinateSystemAxis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordinateSystemAxis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
